+++ sqlite-storage/Cargo.toml
@@ -34,9 +34,10 @@ serde = { version = "1", default-features = false, features = ["derive", "std"] 
 serde_json = "1"
 
 [dev-dependencies]
+storage-tests = { path = "../tests/storage" }
 tempfile = "3"
 time = { version = "0.3", features = ["std"] }
-tokio = { version = "1", features = ["rt", "macros"] }
+tokio = { version = "1", features = ["rt", "rt-multi-thread", "macros", "sync"] }
 
 [build-dependencies]
 regex = "1"
